+++ src/main/java/com/maidclean/springboot/springbootapi/irepository/IEmpregadorRepository.java
@@ -14,6 +14,6 @@ public interface IEmpregadorRepository extends Repository<Empregador, Long>{
  
 	List<Empregador> findAll();
  
-	Empregador findById(Integer id);                  
+	Empregador findById(Long id);                  
 }
